+++ src/Runestone.FSharp.NativeInteropHelper/IMAGE_DATA_DIRECTORY.cs
@@ -4,7 +4,7 @@ using System.Runtime.InteropServices;
 namespace Runestone.FSharp.NativeInteropHelper
 {
     [StructLayout(LayoutKind.Sequential)]
-    public struct IMAGE_DATA_DIRECTORY {
+    internal struct IMAGE_DATA_DIRECTORY {
         public UInt32 VirtualAddress;
         public UInt32 Size;
     }