+++ src/Runestone.FSharp/AssemblyInfo.fs
@@ -6,8 +6,8 @@ open System.Runtime.InteropServices
 [<assembly: AssemblyTitleAttribute("Runestone.FSharp")>]
 [<assembly: AssemblyProductAttribute("Runestone.FSharp")>]
 [<assembly: AssemblyDescriptionAttribute("A tool for analyzing Portable Executable files.")>]
-[<assembly: AssemblyVersionAttribute("0.0.1")>]
-[<assembly: AssemblyFileVersionAttribute("0.0.1")>]
+[<assembly: AssemblyVersionAttribute("0.0.2")>]
+[<assembly: AssemblyFileVersionAttribute("0.0.2")>]
 [<assembly: GuidAttribute("b7390b74-05e0-4eef-9739-88c21fe7b6f9")>]
 do ()
 
@@ -15,6 +15,6 @@ module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "Runestone.FSharp"
     let [<Literal>] AssemblyProduct = "Runestone.FSharp"
     let [<Literal>] AssemblyDescription = "A tool for analyzing Portable Executable files."
-    let [<Literal>] AssemblyVersion = "0.0.1"
-    let [<Literal>] AssemblyFileVersion = "0.0.1"
+    let [<Literal>] AssemblyVersion = "0.0.2"
+    let [<Literal>] AssemblyFileVersion = "0.0.2"
     let [<Literal>] Guid = "b7390b74-05e0-4eef-9739-88c21fe7b6f9"