+++ src/Runestone.FSharp/ExportDirectory.fs
@@ -2,14 +2,16 @@
 
 open System
 
-type ExportedFunction =
-| VirtualAddress
-| 
+type ExportedFunction = {
+    VirtualAddress: uint32
+    Ordinal: uint16
+    Name: string option
+}
 
 type ExportDirectory = {
     Timestamp: DateTime
     Version: Version
     ModuleName: string
-    OrdinalBase: uint16
-
+    OrdinalBase: uint32
+    ExportedFunctions: ExportedFunction []
 }