+++ RELEASE_NOTES.md
@@ -1,2 +1,6 @@
 ### 0.0.1 - August 3 2017
 * Initial release
+
+### 0.0.2 - August 13 2017
+* Added data directories to `OptionalHeader`
+* Added `PortableExecutableFile.ResolveExports`